% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MLE.R
\name{get_MLE}
\alias{get_MLE}
\alias{getMLE}
\title{Print Estimates and Standard Errors from optim Output}
\usage{
get_MLE(opt, dig = 3)

getMLE(opt, dig = 3)
}
\arguments{
\item{opt}{Object resulting from a call to \code{optim}}

\item{dig}{Number of decimal places to use when printing}
}
\value{
Nothing; called for its side effects.
}
\description{
Print Estimates and Standard Errors from optim Output
}
