% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleSummary.R
\name{nimbleSummary}
\alias{nimbleSummary}
\alias{nimble_summary}
\title{Summarize MCMC Samples in an mcmc.list Object Created by NIMBLE}
\usage{
nimbleSummary(samples, params = NULL)

nimble_summary(samples, params = NULL)
}
\arguments{
\item{params}{An optional list of the parameter names used to sort the output}

\item{object}{An object of class \code{mcmc.list}}
}
\value{
A data frame of summary information for each saved parameter
}
\description{
Summarize MCMC Samples in an mcmc.list Object Created by NIMBLE
}
