% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleUI.R
\name{nimbleUI}
\alias{nimbleUI}
\title{Fit a Model with Nimble}
\usage{
nimbleUI(
  code,
  constants,
  inits,
  monitors,
  thin = 1,
  niter = 10000,
  nburnin = 0,
  nchains = 1,
  parallel = FALSE,
  ncores = NULL
)
}
\arguments{
\item{code}{Code expression representing model, output from \code{nimbleCode}}

\item{constants}{Named list of constants in the model}

\item{inits}{A list of initial values or function to generate such a list}

\item{monitors}{A character vector of parameters in the model to monitor}

\item{thin}{Thinning interval for MCMC}

\item{niter}{Number of MCMC iterations to run}

\item{nburnin}{Number of iterations to discard as burn-in}

\item{nchains}{Number of MCMC chains to run}

\item{parallel}{If \code{TRUE}, run MCMC chains in parallel}

\item{ncores}{Manually set number of cores to use when running in parallel.
If \code{NULL} (the default), the number of cores will be the number of chains
or the number of available cores - 1, whichever is smaller.}
}
\value{
An object of class \code{nimbleUI} containing MCMC samples
 as an \code{mcmc.list}, and other info.
}
\description{
This function is a wrapper for fitting a model using Nimble, potentially
running MCMC chains run in parallel. See ?nimble::nimbleMCMC for more details.
}
\seealso{
\code{\link{nimbleMCMC}}
}
