% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleUI.R
\name{print.nimbleUI}
\alias{print.nimbleUI}
\title{Print Summary of nimbleUI Object to the Console}
\usage{
\method{print}{nimbleUI}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{nimbleUI}}

\item{digits}{Number of digits to use when rounding}

\item{...}{Other arguments passed to \code{print}}
}
\value{
Doesn't return anything; called for its side effects.
}
\description{
Print Summary of nimbleUI Object to the Console
}
