% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleUI.R
\name{summary.nimbleUI}
\alias{summary.nimbleUI}
\title{Summarize MCMC Samples in nimbleUI Object}
\usage{
\method{summary}{nimbleUI}(object, ...)
}
\arguments{
\item{object}{An object of class \code{nimbleUI}}

\item{...}{Currently ignored}
}
\value{
A data frame of summary information for each saved parameter
}
\description{
Summarize MCMC Samples in nimbleUI Object
}
