% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_summary.R
\name{tmb_summary}
\alias{tmb_summary}
\title{Summarize Output from TMB}
\usage{
tmb_summary(tmb_object, dig = NULL)
}
\arguments{
\item{tmb_object}{A TMB object created by \code{MakeADFun} that has
been optimized (e.g. with \code{optim})}

\item{dig}{Number of decimal places to use in output}
}
\value{
A matrix of parameter estimates and standard errors.
}
\description{
Summarize Output from TMB
}
