% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbSummary.R
\name{tmbSummary}
\alias{tmbSummary}
\title{Summarize Output from TMB}
\usage{
tmbSummary(tmbObj, dig = NULL)
}
\arguments{
\item{tmbObj}{A TMB object created by \code{MakeADFun} that has
been optimized (e.g. with \code{optim})}

\item{dig}{Number of decimal places to use in output}
}
\value{
A matrix of parameter estimates and standard errors.
}
\description{
Summarize Output from TMB
}
