% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMLE.R
\name{getMLE}
\alias{getMLE}
\alias{get_MLE}
\title{Print Estimates and Standard Errors from optim Output}
\usage{
getMLE(opt, dig = 3)

get_MLE(opt, dig = 3)
}
\arguments{
\item{opt}{Object resulting from a call to \code{optim}}

\item{dig}{Number of decimal places to use when printing}
}
\value{
A matrix of parameter estimates, standard errors, and 95% CI.
}
\description{
Print Estimates and Standard Errors from optim Output
}
