% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleUI.R
\name{nimble_summary}
\alias{nimble_summary}
\alias{summary.nimbleUI}
\title{Summarize MCMC Samples in an mcmc.list Object}
\usage{
nimble_summary(samples, params = NULL)

\method{summary}{nimbleUI}(object, ...)
}
\arguments{
\item{params}{An optional list of the parameter names used to sort the output}

\item{object}{An object of class \code{nimbleUI}}

\item{...}{Currently ignored}
}
\value{
A data frame of summary information for each saved parameter

A data frame of summary information for each saved parameter
}
\description{
Summarize MCMC Samples in an mcmc.list Object

Summarize MCMC Samples in nimbleUI Object
}
